+++ app/Http/Controllers/CashRegisterOperationController.php
@@ -48,7 +48,7 @@ class CashRegisterOperationController extends Controller
             $cashRegisterOperation = CashRegisterOperation::create($request->all());
 
             $this->updateStaffStatusBarInfo(
-                (int) $request->input('amount'),
+                (float) $request->input('amount'),
                 (string) ($cashRegisterOperation->cash_register_operation_type->state == 1) ? '+' : '-'
             );
